Refactor: simplify generator.

+++ src/day1.rs
@@ -2,16 +2,17 @@ use std::num::ParseIntError;
 
 use aoc_runner_derive::{aoc, aoc_generator};
 use eyre::Result;
-use nom::character::complete::{i16, one_of};
+use nom::character::complete::{one_of, usize};
 use nom::combinator::map_res;
 use nom::{IResult, Parser};
 
 #[derive(Copy, Clone, Debug)]
 enum Rotation {
-    Left(i16),
-    Right(i16),
+    Left(usize),
+    Right(usize),
 }
 
+#[derive(Clone, Debug)]
 struct Input1 {
     pub data: Vec<Rotation>,
 }
@@ -26,13 +27,13 @@ fn input_generator(input: &str) -> Result<Input1> {
 }
 
 #[aoc(day1, part1)]
-fn solve_part1(input: &Input1) -> u32 {
-    let mut cnt = 0;
+fn solve_part1(input: &Input1) -> usize {
+    let mut cnt = 0usize;
 
     input.data.iter().fold(50, |acc, r| {
         let m = match r {
-            Rotation::Left(amount) => -(*amount),
-            Rotation::Right(amount) => *amount,
+            Rotation::Left(amount) => -(*amount as i32),
+            Rotation::Right(amount) => *amount as i32,
         };
         let res = (acc + m) % 100;
         if res == 0 {
@@ -44,7 +45,7 @@ fn solve_part1(input: &Input1) -> u32 {
 }
 
 fn parse_rotation(input: &str) -> IResult<&str, Rotation> {
-    let r = |(dir, amount): (char, std::primitive::i16)| -> Result<Rotation, ParseIntError> {
+    let r = |(dir, amount): (char, std::primitive::usize)| -> Result<Rotation, ParseIntError> {
         match dir {
             'L' => Ok(Rotation::Left(amount)),
             'R' => Ok(Rotation::Right(amount)),
@@ -54,5 +55,5 @@ fn parse_rotation(input: &str) -> IResult<&str, Rotation> {
             ),
         }
     };
-    map_res((one_of("LR"), i16), r).parse(input)
+    map_res((one_of("LR"), usize), r).parse(input)
 }
